feat: Add NumberOfPages component
