+++ packages/page-navigation/src/types/GoToPage.ts
@@ -1,0 +1,24 @@
+/**
+ * A React component to view a PDF document
+ *
+ * @see https://react-pdf-viewer.dev
+ * @license https://react-pdf-viewer.dev/license
+ * @copyright 2019-2022 Nguyen Huu Phuoc <me@phuoc.ng>
+ */
+
+import * as React from 'react';
+
+export interface GoToPageMenuItemProps {
+    onClick: () => void;
+}
+
+export interface RenderGoToPageProps {
+    isDisabled: boolean;
+    onClick: () => void;
+}
+
+export type RenderGoToPage = (props: RenderGoToPageProps) => React.ReactElement;
+
+export interface GoToPageProps {
+    children?: RenderGoToPage;
+}
