+++ packages/page-navigation/src/types/index.ts
@@ -6,19 +6,4 @@
  * @copyright 2019-2022 Nguyen Huu Phuoc <me@phuoc.ng>
  */
 
-import * as React from 'react';
-
-export interface GoToPageMenuItemProps {
-    onClick: () => void;
-}
-
-export interface RenderGoToPageProps {
-    isDisabled: boolean;
-    onClick: () => void;
-}
-
-export type RenderGoToPage = (props: RenderGoToPageProps) => React.ReactElement;
-
-export interface GoToPageProps {
-    children?: RenderGoToPage;
-}
+export * from './GoToPage';
